Fixed a lot of stuff, removed dependency on the weird google thing
